(deploy reorg/refactor) Relocates all JS-centric code to deployment/javascript.js
Still more granular work to do, but this gets the split in.
