+++ lib/tessel/commands.js
@@ -60,7 +60,7 @@ module.exports.appendStdinToFile = function(filepath) {
   return ['tee', '-a', filepath];
 };
 module.exports.chmod = function(permission, filepath) {
-  return ['chmod', '+x', filepath];
+  return ['chmod', permission, filepath];
 };
 module.exports.connectedNetworkStatus = function() {
   return ['ubus', 'call', 'network.interface.lan', 'status'];
