+++ lib/tessel/deployment/index.js
@@ -9,7 +9,7 @@ var exportables = {
     input = String(input).toLowerCase();
     for (var key in languages) {
       if (languages[key] &&
-        (input === languages[key].lang.name || input === languages[key].lang.extname)) {
+        (input === languages[key].meta.name || input === languages[key].meta.extname)) {
         return languages[key];
       }
     }
