+++ lib/tessel/deployment/python.js
@@ -1,5 +1,5 @@
 var exportables = {
-  lang: {
+  meta: {
     name: 'python',
     extname: 'py',
     binary: 'python',
