+++ lib/tessel/deployment/rust.js
@@ -1,5 +1,5 @@
 var exportables = {
-  lang: {
+  meta: {
     name: 'rust',
     extname: 'rs',
     binary: '',
