+++ lib/tessel/erase.js
@@ -22,7 +22,7 @@ Tessel.prototype.eraseScript = function() {
         });
     })
     .catch((error) => {
-      // If we get a notice that the command failed
+      // If we ge,t a notice that the command failed
       if (error.message.includes('Command failed')) {
         // Let the user know what went wrong
         return Promise.reject('No files have been pushed. Run `t2 push FILE` to push to Flash.');
