+++ test/common/bootstrap.js
@@ -42,7 +42,6 @@ global.Tessel = require('../../lib/tessel/tessel');
 global.commands = require('../../lib/tessel/commands');
 global.deploy = require('../../lib/tessel/deploy');
 global.deployment = require('../../lib/tessel/deployment/index');
-global.deployLists = require('../../lib/tessel/deploy-lists');
 global.provision = require('../../lib/tessel/provision');
 global.RSA = require('../../lib/tessel/rsa-delegation');
